+++ components/home/Feed.tsx
@@ -1,15 +1,10 @@
 // ... 기존 import 문 ...
-import { useState } from "react";
 import { UserAvatarWithLabel } from "../my";
 import Image from "next/image";
+import AudioModule from "./AudioModule";
+import EmojiSelector from "./EmojiSelector";
 
 const Feed: React.FC = () => {
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-
-  const toggleMenu = () => {
-    setIsMenuOpen(!isMenuOpen);
-  };
-
   return (
     <div className="flex-col w-full max-w-md border-b border-gray-200">
       <UserAvatarWithLabel
@@ -27,31 +22,8 @@ const Feed: React.FC = () => {
           style={{ width: "100%", height: "auto" }}
           className="rounded-md shadow-md"
         />
-        <button
-          onClick={toggleMenu}
-          className="absolute bottom-2 left-2 bg-gray-100 text-gray-800 text-xs font-medium px-1 py-1 rounded-full dark:bg-gray-700 dark:text-gray-300 opacity-75"
-        >
-          😄
-        </button>
-        {isMenuOpen && (
-          <div className="absolute bottom-2 left-10 bg-white rounded-full shadow-lg flex space-x-2 opacity-75">
-            <button className="block w-full text-left px-1 y-2 hover:bg-gray-100">
-              😀
-            </button>
-            <button className="block w-full text-left px-1 y-2 hover:bg-gray-100">
-              😀
-            </button>
-            <button className="block w-full text-left px-1 y-2 hover:bg-gray-100">
-              😀
-            </button>
-            <button className="block w-full text-left px-1 y-2 hover:bg-gray-100">
-              😀
-            </button>
-            <button className="block w-full text-left px-1 y-2 hover:bg-gray-100 pr-4">
-              😀
-            </button>
-          </div>
-        )}
+        <AudioModule src="https://codeskulptor-demos.commondatastorage.googleapis.com/GalaxyInvaders/theme_01.mp3" />
+        <EmojiSelector />
       </div>
       <div className="mb-10">
         <p className="w-full border-none text-sm bg-white font-mono m-2">
